Consolidate lightningd and lightningserver.
Sacrifices daemonization capabilities, but that was currently unused.
