+++ requirements.txt
@@ -1,6 +1,8 @@
 Flask==0.10.1
+Flask-SQLAlchemy==2.0
 Jinja2==2.7.3
 MarkupSafe==0.23
+SQLAlchemy==1.0.8
 Werkzeug==0.10.4
 astroid==1.3.6
 blinker==1.4
